fix: diary core bugs + simplify home UI
Bug fixes (P0):
- Fix TagEditor color truncation (Long.toInt() -> Long)
- Fix search history never showing (contradictory conditions)
- Fix streak calculation to start from most recent entry
- Fix share text to use plainText instead of regex HTML stripping

Bug fixes (P1):
- MoodSlider now supports deselection (click again to clear)
- Calendar prevents scrolling to future months
- Editor rejects saving empty entries
- FAB gradient uses correct dark mode colors
- Weather field comment matches actual values

UI simplification:
- Home: Stats card -> compact single-line row
- Home: Calendar collapsed by default, toggle button added
- Home: Calendar shows selected date with clear button

+++ app/src/main/java/com/diary/app/ui/editor/TagEditor.kt
@@ -57,7 +57,7 @@ fun TagEditor(
     ) {
         allTags.forEach { tag ->
             val isSelected = tag.id in selectedTagIds
-            val tagColor = Color(tag.color.toInt())
+            val tagColor = Color(tag.color)
 
             TagChip(
                 tagName = tag.name,
